refactor: clean up more files
