+++ wallet-sdk-ui/src/iosMain/kotlin/com/michaeltchuang/walletsdk/ui/liquidAuth/service/LiquidAuthConnectionManager.kt
@@ -248,6 +248,7 @@ class IOSLiquidAuthConnectionManager : LiquidAuthConnectionManager {
             IOSLiquidStreamCreator(
                 mppServerConfig = mppServerConfig,
                 serverConfig = serverConfig,
+                getRemainingSessionVaultBalanceUseCase = getRemainingBalanceUseCase,
             )
 
         creator.onSessionStarted = { sid ->
